+++ packages/cater-build/app/provider.js
@@ -1,10 +1,24 @@
 // Copyright Jon Williams 2017. See LICENSE file.
 import React from 'react';
 
-class EmptyProvider extends React.Component {
+class Provider extends React.Component {
   render() {
-    return this.props.children;
+    let { children } = this.props;
+
+    /* eslint-disable no-undef */
+    if (
+      (typeof MODE !== 'undefined' && MODE === 'dev') ||
+      (process && process.env && process.env.CATER_MODE === 'dev')
+    ) {
+      const script =
+        'window.onerror = function() { console.log(arguments); window.errorState = arguments; }';
+      children = [
+        <script key="script" type="text/javascript" dangerouslySetInnerHTML={{ __html: script }} />
+      ].concat(this.props.children);
+    }
+
+    return children;
   }
 }
 
-export default EmptyProvider;
+export default Provider;