+++ packages/cater-build/src/config-default.js
@@ -22,6 +22,7 @@ const buildDefaults = {
   plugins: 'auto',
   publicPath: '/static/',
   sideNames: ['client', 'server'],
+  startOnError: true,
   staticDirectory: 'static',
   universalNames: ['app'],
   universalPrefix: 'app'